+++ tests/node.rs
@@ -1,39 +1,19 @@
-use once_cell::sync::Lazy;
-
+use hapi_rs::session::CookResult;
 use hapi_rs::{
     node::{
         HoudiniNode, KeyFrame, NodeFlags, NodeType, PresetType, RSTOrder, StatusVerbosity,
         TransformComponent, TransformEuler,
     },
     parameter::Parameter,
-    session::{quick_session, CookResult, Session, SessionOptions},
     Result,
 };
 
-thread_local! {
-static SESSION: Lazy<Session> = Lazy::new(|| {
-    env_logger::try_init().ok();
-    let opt = SessionOptions::builder().threaded(true).build();
-    let session = quick_session(Some(&opt)).expect("Could not create test session");
-    session
-        .load_asset_file("otls/hapi_geo.hda")
-        .expect("load asset");
-    session
-        .load_asset_file("otls/hapi_vol.hda")
-        .expect("load asset");
-    session
-        .load_asset_file("otls/hapi_parms.hda")
-        .expect("load asset");
-    session
-        .load_asset_file("otls/sesi/SideFX_spaceship.hda")
-        .expect("load asset");
-    session
-});
-}
+mod _utils;
+use _utils::*;
 
 #[test]
 fn node_create() {
-    SESSION.with(|session| {
+    with_session(|session| {
         let node = session.create_node("Object/spaceship").unwrap();
         assert_eq!(
             node.cook_count(NodeType::None, NodeFlags::None, true)
@@ -47,12 +27,14 @@ fn node_create() {
             1
         );
         assert!(matches!(session.cook().unwrap(), CookResult::Succeeded));
+        Ok(())
     })
+    .unwrap()
 }
 
 #[test]
 fn node_set_node_display_flag() -> Result<()> {
-    SESSION.with(|session| {
+    with_session(|session| {
         let sop = session.create_node("Object/geo").unwrap();
         let sphere = session.node_builder("sphere").with_parent(&sop).create()?;
         let cube = session.node_builder("box").with_parent(&sop).create()?;
@@ -69,7 +51,7 @@ fn node_set_node_display_flag() -> Result<()> {
 
 #[test]
 fn node_inputs_and_outputs() {
-    SESSION.with(|session| {
+    with_session(|session| {
         let node = session.create_node("Object/hapi_geo").unwrap();
         let geo = node.geometry().unwrap().unwrap();
         let mut input = geo.node.input_node(0).unwrap();
@@ -87,12 +69,14 @@ fn node_inputs_and_outputs() {
             n.get_input_name(0).unwrap(),
             "Geometry to Process with Wrangle"
         );
+        Ok(())
     })
+    .unwrap()
 }
 
 #[test]
 fn node_search() {
-    SESSION.with(|session| {
+    with_session(|session| {
         let asset = session.create_node("Object/hapi_geo").unwrap();
         asset.cook_blocking().unwrap();
         let nope = asset.get_child_by_path("bla").unwrap();
@@ -108,13 +92,15 @@ fn node_search() {
         let children = asset
             .find_children_by_type(NodeType::Sop, NodeFlags::Display, true)
             .unwrap();
-        assert_ne!(children.len(), 0)
+        assert_ne!(children.len(), 0);
+        Ok(())
     })
+    .unwrap()
 }
 
 #[test]
 fn node_transform() {
-    SESSION.with(|session| {
+    with_session(|session| {
         let obj = session.create_node("Object/null").unwrap();
         let t = obj.get_transform(None, None).unwrap();
         assert_eq!(t.position(), [0.0, 0.0, 0.0]);
@@ -133,12 +119,14 @@ fn node_transform() {
         session
             .get_composed_object_transform(obj.parent_node().unwrap(), RSTOrder::Default)
             .unwrap();
+        Ok(())
     })
+    .unwrap()
 }
 
 #[test]
 fn node_save_and_load() {
-    SESSION.with(|session| {
+    with_session(|session| {
         let cam = session.create_node("Object/cam").unwrap();
         let tmp = std::env::temp_dir().join("node");
         cam.save_to_file(&tmp).expect("save_to_file");
@@ -147,22 +135,26 @@ fn node_save_and_load() {
         std::fs::remove_file(&tmp).unwrap();
         cam.delete().unwrap();
         new.delete().unwrap();
+        Ok(())
     })
+    .unwrap()
 }
 
 #[test]
 fn node_number_of_geo_outputs() {
-    SESSION.with(|session| {
+    with_session(|session| {
         let node = session.create_node("Object/hapi_geo").unwrap();
         assert_eq!(node.number_of_geo_outputs().unwrap(), 2);
         let infos = node.geometry_output_nodes().unwrap();
         assert_eq!(infos.len(), 2);
+        Ok(())
     })
+    .unwrap()
 }
 
 #[test]
 fn node_get_message_nodes() {
-    SESSION.with(|session| {
+    with_session(|session| {
         let node = session.create_node("Object/hapi_geo").unwrap();
         node.cook_blocking().unwrap();
         let message_nodes = node.get_message_nodes().unwrap();
@@ -174,29 +166,35 @@ fn node_get_message_nodes() {
             .get_cook_result_string(StatusVerbosity::Statusverbosity2)
             .unwrap();
         assert!(msg.contains("Warning generated by Python node"));
+        Ok(())
     })
+    .unwrap()
 }
 
 #[test]
 fn node_output_names() {
-    SESSION.with(|session| {
+    with_session(|session| {
         let node = session.create_node("Object/hapi_parms").unwrap();
         let outputs = node.get_output_names().unwrap();
         assert_eq!(outputs[0], "nothing");
+        Ok(())
     })
+    .unwrap()
 }
 
 #[test]
 fn node_get_parm_with_tag() {
-    SESSION.with(|session| {
+    with_session(|session| {
         let node = session.create_node("Object/hapi_parms").unwrap();
         assert!(node.parameter_with_tag("my_tag").unwrap().is_some());
+        Ok(())
     })
+    .unwrap()
 }
 
 #[test]
 fn node_set_animate_transform() {
-    SESSION.with(|session| {
+    with_session(|session| {
         let bone = session.create_node("Object/bone").unwrap();
         let ty = [
             KeyFrame {
@@ -218,12 +216,14 @@ fn node_set_animate_transform() {
         if let Parameter::Float(p) = bone.parameter("ty").unwrap() {
             assert_eq!(p.get(1).unwrap(), 5.0);
         }
+        Ok(())
     })
+    .unwrap()
 }
 
 #[test]
 fn node_get_set_preset() {
-    SESSION.with(|session| {
+    with_session(|session| {
         let node = session.create_node("Object/null").unwrap();
         if let Parameter::Float(p) = node.parameter("scale").unwrap() {
             assert_eq!(p.get(0).unwrap(), 1.0);
@@ -232,5 +232,7 @@ fn node_get_set_preset() {
             node.set_preset("test", PresetType::Binary, &save).unwrap();
             assert_eq!(p.get(0).unwrap(), 1.0);
         }
+        Ok(())
     })
+    .unwrap()
 }
